+++ CHANGELOG.md
@@ -14,6 +14,11 @@ Prefix your message with one of the following:
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+- [Fixed] Handle BigNumber raising when receiving invalid numbers by default on
+  v10.
+
 ## v4.5.2 - Feb 5, 2026
 
 - [Fixed] Fix `numberToHuman` separator not applied when